requirements.txt
asgiref==3.4.1
Django==3.2.9
django-taggit==2.0.0
importlib-metadata==4.8.2
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
zipp==3.6.0
